refactor(config): split config params into two files
